refactor(caddy_tls): collapse Otto::Services umbrella into Otto::CaddyTLS
Otto::Services was introduced (#175) as a home for network-service
integrations, on the assumption that CSP reporting (#174) would be its
second tenant. CSP instead shipped as Otto::Security::CSP -- it is the
receiving half of Otto's own CSP support and belongs beside the emitting
half, not in a generic bucket. With its presumed second tenant placed
elsewhere, the umbrella had a single occupant.

Collapse it to a feature-named, top-level Otto::CaddyTLS, matching the
Otto::MCP precedent for external-system integrations. The public API
(enable_caddy_tls!, caddy_tls_enabled?) is unchanged; this is an internal
rename plus doc/changelog updates that record the decision.

- lib/otto/services/* -> lib/otto/caddy_tls/*; Server/PermissionHandler
  de-nest one level; LocalhostGuard scoped under CaddyTLS (promotable if a
  second internal-only integration ever needs it)
- Track docs/reverse-proxy-network-services.md (allowlisted in
  docs/.gitignore like the sibling design docs, which fixes the changelog's
  reference to it) and revise it to record the namespace decision
- Full suite green (1372 examples, 0 failures); rubocop clean on lib
