+++ changelog.d/20260701_000000_claude_issue-175.rst
@@ -1,8 +1,8 @@
 Added
 -----
 
-- ``Otto::Services``, a modular home for network-service integrations, with the
-  Caddy on-demand TLS permission endpoint as the first one. Enable it with
+- ``Otto::CaddyTLS``, an opt-in integration for Caddy's on-demand TLS permission
+  endpoint (a feature-named module alongside ``Otto::MCP``). Enable it with
   ``otto.enable_caddy_tls! { |domain| ... }`` — the block decides (truthy allows,
   falsy/nil/raise denies). One endpoint serves both Caddy's deprecated ``ask``
   directive and the ``permission http`` module. (delano/otto#175)
@@ -11,7 +11,7 @@ Security
 --------
 
 - The Caddy permission endpoint is loopback-only by default via
-  ``Otto::Services::LocalhostGuard``: allowed only when the raw TCP peer is
+  ``Otto::CaddyTLS::LocalhostGuard``: allowed only when the raw TCP peer is
   loopback and the request carries no forwarding header. Fail-closed on missing
   block, blank domain, and callback errors.
