+++ docs/.gitignore
@@ -5,3 +5,4 @@
 !ipaddr-encoding-quirk.md
 !modern-authentication-authorization-landscape.md
 !multi-strategy-authentication-design.md
+!reverse-proxy-network-services.md
