+++ lib/otto/caddy_tls.rb
@@ -1,0 +1,7 @@
+# lib/otto/caddy_tls.rb
+#
+# frozen_string_literal: true
+
+require_relative 'caddy_tls/core'
+require_relative 'caddy_tls/localhost_guard'
+require_relative 'caddy_tls/server'
