--- lib/otto/services.rb
@@ -1,7 +1,0 @@
-# lib/otto/services.rb
-#
-# frozen_string_literal: true
-
-require_relative 'services/core'
-require_relative 'services/localhost_guard'
-require_relative 'services/caddy_tls'
