--- lib/otto/services/caddy_tls.rb
@@ -1,5 +1,0 @@
-# lib/otto/services/caddy_tls.rb
-#
-# frozen_string_literal: true
-
-require_relative 'caddy_tls/server'
